+++ CHANGELOG.md
@@ -3,14 +3,17 @@
 **0.8.1.0** — Snapshot release
 
 * Improve documentation and handling of call buffering (`CallBufferMode`). The default call buffer mode was documented incorrectly, it is `BufferRun`. [#163][], [#191][], [#192][]
+* Add support for [custom DOM events][customevent] (`CustomEvent`). [#196][]
 * Expose JavaScript FFI functions `toJSObject` and `liftJSWindow` in `Graphics.UI.Threepenny`. This is useful for linking the lifetime of JavaScript objects to the lifetime of `Element`. [#181][]
 * Use `jsLog` parameter to log exceptions. [#185][]
 
+  [customevent]: https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent/CustomEvent
   [#163]: https://github.com/HeinrichApfelmus/threepenny-gui/issues/163
   [#181]: https://github.com/HeinrichApfelmus/threepenny-gui/issues/181
   [#185]: https://github.com/HeinrichApfelmus/threepenny-gui/issues/185
   [#191]: https://github.com/HeinrichApfelmus/threepenny-gui/issues/191
   [#192]: https://github.com/HeinrichApfelmus/threepenny-gui/issues/192
+  [#196]: https://github.com/HeinrichApfelmus/threepenny-gui/issues/196
 
 
 **0.8.0.1** — Maintenance release